Convert middleware to pipeline with batch processing support
Problem with current implementation is that if by the end of the
crawl there are remaining items that were not sent to Monkeylearn API
yet - they will never be sent and will never be collected. E.g.
if `MONKEYLEARN_BATCH_SIZE` is set to 200 and only 100 items are
collected by the spider - all of the items will stuck in that
middleware and will never be passed to pipelines.

This commit replaces middleware with a pipeline that processes items
in batches. Pipeline takes items, stores it in the batch and returns
a Deferred object. Monkeylearn API requester is running is a separate
thread. Once `MONKEYLEARN_BATCH_SIZE` items are collected or an
internal timeout occurs - items are sent to the Monkeylearn API and
on response deferreds are activated and items are sent to the next
pipeline.

- Removed scrapy_monkeylearn.middlewares.MonkeylearnMiddleware
- Added scrapy_monkeylearn.pipelines.MonkeylearnPipeline
- Added missing setup.py
- Updated requrements.txt
- Added Python 3 support
- Updated README

Some of the changes are taken from https://github.com/scrapy-plugins/scrapy-monkeylearn/commit/4e337af9c2803cfcff1ef720d62e048b868f0d82